Added drop down menu
Drop down menu works. I just put print statements as filler, but we can put code into them as we go. also wasn't sure if we needed clipboard/joy stick/ sound, but should be easy to add now if needed. One problem is the drop down menu doesn't show up on the game screen... Messed around with it, but ultimately wanted it to show up in the upper left corner.